chore(SwingUtils): removed implicit conversion on option string and int
