+++ src/main/scala/ulisse/infrastructures/view/components/ExtendedSwing.scala
@@ -1,5 +1,7 @@
 package ulisse.infrastructures.view.components
 
+import ulisse.infrastructures.view.common.Themes.Theme
+import ulisse.infrastructures.view.components.composed.ComposedImageLabel
 import ulisse.infrastructures.view.components.decorators.ImageEffects.{PictureEffect, SVGEffect}
 import ulisse.infrastructures.view.components.decorators.SwingEnhancements.{EnhancedLook, FontEffect, ShapeEffect}
 import ulisse.infrastructures.view.components.styles.Styles
@@ -66,3 +68,43 @@ object ExtendedSwing:
   case class SLabel(label: String) extends Label(label) with ShapeEffect with FontEffect
 
   case class STextField(colum: Int) extends TextField(colum) with ShapeEffect with FontEffect
+
+  extension (panes: Map[ComposedImageLabel, Component])
+    /** Returns [[JTabbedPane]] given `panes`, a mapping of tab item [[ComposedImageLabel]] to theirs views. */
+    def toTabbedPane: JTabbedPane =
+      val tabbedPane = ComposedSwing.createTabbedPane(panes.keys.toList: _*)
+      panes.foreach((k, p) => tabbedPane.paneOf(k).contents += p)
+      tabbedPane
+
+  /** A specific [[STextField]] that accept only numbers. */
+  class SNumberField(cols: Int) extends STextField(cols):
+
+    import ulisse.infrastructures.view.common.Themes.withAlpha
+
+    private val textFieldPadding    = Styles.createPadding(10, 5)
+    private val textFieldBackground = Theme.light.background.withAlpha(50)
+
+    rect = rect.withPadding(textFieldPadding)
+    rectPalette = rectPalette.withBackground(textFieldBackground)
+    import scala.swing.event.ValueChanged
+    reactions += {
+      case ValueChanged(_) => Swing.onEDT(if (!text.matches("^[0-9]*$")) text = text.filter(_.isDigit))
+    }
+
+  import ulisse.infrastructures.view.components.composed.ComposedSwing
+  import ulisse.infrastructures.view.common.Themes.withAlpha
+  import ulisse.infrastructures.view.components.ExtendedSwing
+
+  /** A themed field label for given `fieldComponent` */
+  class SFieldLabel(text: String)(fieldComponent: Component) extends ComposedSwing:
+    private val fieldBackground = Theme.light.background.withAlpha(50)
+    private val mainPanel       = ExtendedSwing.SBoxPanel(Orientation.Vertical)
+    private val labelPanel      = ExtendedSwing.SFlowPanel()
+    private val label           = ExtendedSwing.SLabel(text)
+
+    fieldComponent.background = fieldBackground
+    labelPanel.contents += label
+    mainPanel.contents += labelPanel
+    mainPanel.contents += fieldComponent
+
+    override def component[T >: Component]: T = mainPanel
