+++ src/main/scala/ulisse/infrastructures/view/utils/SwingUtils.scala
@@ -6,7 +6,6 @@ import ulisse.infrastructures.view.components.composed.{ComposedImageLabel, Comp
 import ulisse.infrastructures.view.components.composed.ComposedSwing.JTabbedPane
 import ulisse.infrastructures.view.components.styles.Styles
 
-import scala.swing.event.ValueChanged
 import scala.swing.*
 
 object SwingUtils:
@@ -15,16 +14,11 @@ object SwingUtils:
   val labelFont = new Font("Arial", java.awt.Font.BOLD, 14)
   val valueFont = new Font("Arial", java.awt.Font.PLAIN, 14)
 
-  @SuppressWarnings (Array ("org.wartremover.warts.ImplicitConversion"))
   extension (s: Option[String])
-    implicit def defaultString: String =
-      s.getOrElse("N/A")
+    def defaultString: String = s.getOrElse("N/A")
 
   extension (i: Option[Int])
-    implicit def defaultIntString: String =
-      i match
-        case Some(v) => v.toString
-        case None    => "N/A"
+    def defaultIntString: String = i.map(_.toString).defaultString
 
   extension (text: String)
     def headerLabel: Label =
@@ -47,40 +41,3 @@ object SwingUtils:
         contents = mainPanel
         visible = true
       }
-
-  extension (panes: Map[ComposedImageLabel, Component])
-    /** Returns [[JTabbedPane]] given `panes`, a mapping of tab item [[ComposedImageLabel]] to theirs views. */
-    def toTabbedPane: JTabbedPane =
-      val tabbedPane = ComposedSwing.createTabbedPane(panes.keys.toList: _*)
-      panes.foreach((k, p) => tabbedPane.paneOf(k).contents += p)
-      tabbedPane
-
-  /** A specific [[STextField]] that accept only numbers. */
-  class SNumberField(cols: Int) extends STextField(cols):
-    import ulisse.infrastructures.view.common.Themes.withAlpha
-    private val textFieldPadding    = Styles.createPadding(10, 5)
-    private val textFieldBackground = Theme.light.background.withAlpha(50)
-
-    rect = rect.withPadding(textFieldPadding)
-    rectPalette = rectPalette.withBackground(textFieldBackground)
-    reactions += {
-      case ValueChanged(_) => Swing.onEDT(if (!text.matches("^[0-9]*$")) text = text.filter(_.isDigit))
-    }
-
-  import ulisse.infrastructures.view.components.composed.ComposedSwing
-  import ulisse.infrastructures.view.common.Themes.withAlpha
-  import ulisse.infrastructures.view.components.ExtendedSwing
-
-  /** A themed field label for given `fieldComponent` */
-  class SFieldLabel(text: String)(fieldComponent: Component) extends ComposedSwing:
-    private val fieldBackground = Theme.light.background.withAlpha(50)
-    private val mainPanel       = ExtendedSwing.SBoxPanel(Orientation.Vertical)
-    private val labelPanel      = ExtendedSwing.SFlowPanel()
-    private val label           = ExtendedSwing.SLabel(text)
-
-    fieldComponent.background = fieldBackground
-    labelPanel.contents += label
-    mainPanel.contents += labelPanel
-    mainPanel.contents += fieldComponent
-
-    override def component[T >: Component]: T = mainPanel
